wlan1-manager/requirements.txt
geopy
pyyaml
RPi.GPIO; platform_machine == "armv7l" or platform_machine == "aarch64"
